+++ packages/theme/index.ts
@@ -1,8 +1,8 @@
 import * as Extend from "./extend";
 import * as Plugins from "./plugins";
 
-export const aristobyteui = () => Object.values(Plugins);
+const aristobyteui = { plugin: Plugins.plugin, extend: Extend.extend };
 
 export * from "./extend";
 export * from "./plugins";
-export default { plugin: aristobyteui, extend: Extend.extend };
+export { aristobyteui, aristobyteui as default };