+++ packages/theme/plugins/spinner.ts
@@ -1,8 +1,17 @@
 import plugin from "tailwindcss/plugin";
-import type { ColorsType, ParsedVariantsType, VariantsType } from "./types";
+import type {
+  ColorsType,
+  ParsedVariantsType,
+  PluginType,
+  VariantsType,
+} from "./types";
 import { themeToMapping } from "../utilities";
 
-export const spinner = plugin(function ({ addComponents, addBase, theme }) {
+export const spinner: PluginType = plugin(function ({
+  addComponents,
+  addBase,
+  theme,
+}) {
   const colors = theme("colors") as ColorsType;
   const variants = themeToMapping<VariantsType, ParsedVariantsType>(
     theme("variants")