+++ packages/message-box/CHANGELOG.md
@@ -1,5 +1,23 @@
 # @aristobyte-ui/message-box
 
+## 1.0.72
+
+### Patch Changes
+
+- feat: add plugin type in packages/theme
+
+- Updated dependencies []:
+  - @aristobyte-ui/utils@1.0.72
+
+## 1.0.71
+
+### Patch Changes
+
+- feat: resolve type issue in packages/theme and rename the exports
+
+- Updated dependencies []:
+  - @aristobyte-ui/utils@1.0.71
+
 ## 1.0.70
 
 ### Patch Changes