+++ packages/react/CHANGELOG.md
@@ -1,5 +1,39 @@
 # @aristobyte-ui/react
 
+## 1.0.72
+
+### Patch Changes
+
+- feat: add plugin type in packages/theme
+
+- Updated dependencies []:
+  - @aristobyte-ui/message-box@1.0.72
+  - @aristobyte-ui/dropdown@1.0.72
+  - @aristobyte-ui/presets@1.0.72
+  - @aristobyte-ui/spinner@1.0.72
+  - @aristobyte-ui/anchor@1.0.72
+  - @aristobyte-ui/button@1.0.72
+  - @aristobyte-ui/switch@1.0.72
+  - @aristobyte-ui/radio@1.0.72
+  - @aristobyte-ui/utils@1.0.72
+
+## 1.0.71
+
+### Patch Changes
+
+- feat: resolve type issue in packages/theme and rename the exports
+
+- Updated dependencies []:
+  - @aristobyte-ui/anchor@1.0.71
+  - @aristobyte-ui/button@1.0.71
+  - @aristobyte-ui/dropdown@1.0.71
+  - @aristobyte-ui/message-box@1.0.71
+  - @aristobyte-ui/presets@1.0.71
+  - @aristobyte-ui/radio@1.0.71
+  - @aristobyte-ui/spinner@1.0.71
+  - @aristobyte-ui/switch@1.0.71
+  - @aristobyte-ui/utils@1.0.71
+
 ## 1.0.70
 
 ### Patch Changes